// front/src/app/models/enums/rango.enum.ts
export enum Rango {
  INTERN = 'INTERN',
  TRAINEE = 'TRAINEE',
  ASSOCIATE = 'ASSOCIATE',
  JUNIOR = 'JUNIOR',
  SENIOR = 'SENIOR',
  LEAD = 'LEAD',
  SPECIALIST = 'SPECIALIST',
  COORDINATOR = 'COORDINATOR',
  MANAGER = 'MANAGER',
  PRINCIPAL = 'PRINCIPAL',
  CONSULTANT = 'CONSULTANT',
  DIRECTOR = 'DIRECTOR',
  VICE_PRESIDENT = 'VICE_PRESIDENT',
  EXECUTIVE = 'EXECUTIVE',
  CHIEF = 'CHIEF',
  PRESIDENT = 'PRESIDENT',
  PARTNER = 'PARTNER',
  FELLOW = 'FELLOW'
}